requirements.txt
aniso8601==9.0.1
asgiref==3.3.4
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.1
cloudinary==1.29.0
colorama==0.4.4
distlib==0.3.2
Django==3.2
djangorestframework==3.12.4
dnspython==2.1.0
email-validator==1.1.3
filelock==3.0.12
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.19.3
package-name==0.1
Pillow==9.0.1
pipenv==2021.5.29
platformdirs==2.3.0
psycopg2==2.9.3
pycparser==2.21
pytz==2021.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
spark-parser==1.8.9
SQLAlchemy==1.4.24
sqlparse==0.4.1
typing-extensions==3.7.4.3
uncompyle6==3.8.0
urllib3==1.26.8
virtualenv==20.7.2
virtualenv-clone==0.5.6
Werkzeug==2.0.2
WTForms==2.3.3
xdis==6.0.3
zipp==3.5.0
